ML-client/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.26.0
spotipy==2.23.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
google-genai